Updated the wrong paths after renaming and spliting the api and actions file
