+++ cm-client/src/main/resources/components/SkillItem.js
@@ -2,7 +2,7 @@
  * Created by blajv on 22.06.2017.
  */
 import React from 'react';
-import { fetchTagForCandidateSkill} from "../utils/CandidatesAPI";
+import { fetchTagForCandidateSkill} from "../utils/CandidateAPI";
 /**
  * Component which describes a single set of skills a candidate has
  */
